src/main/webapp/app/entities/sach/route/sach-routing-resolve.service.spec.ts
jest.mock('@angular/router');

import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, Router } from '@angular/router';
import { of } from 'rxjs';

import { ISach, Sach } from '../sach.model';
import { SachService } from '../service/sach.service';

import { SachRoutingResolveService } from './sach-routing-resolve.service';

describe('Service Tests', () => {
  describe('Sach routing resolve service', () => {
    let mockRouter: Router;
    let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
    let routingResolveService: SachRoutingResolveService;
    let service: SachService;
    let resultSach: ISach | undefined;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule],
        providers: [Router, ActivatedRouteSnapshot],
      });
      mockRouter = TestBed.inject(Router);
      mockActivatedRouteSnapshot = TestBed.inject(ActivatedRouteSnapshot);
      routingResolveService = TestBed.inject(SachRoutingResolveService);
      service = TestBed.inject(SachService);
      resultSach = undefined;
    });

    describe('resolve', () => {
      it('should return ISach returned by find', () => {
        // GIVEN
        service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
        mockActivatedRouteSnapshot.params = { id: 123 };

        // WHEN
        routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
          resultSach = result;
        });

        // THEN
        expect(service.find).toBeCalledWith(123);
        expect(resultSach).toEqual({ id: 123 });
      });

      it('should return new ISach if id is not provided', () => {
        // GIVEN
        service.find = jest.fn();
        mockActivatedRouteSnapshot.params = {};

        // WHEN
        routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
          resultSach = result;
        });

        // THEN
        expect(service.find).not.toBeCalled();
        expect(resultSach).toEqual(new Sach());
      });

      it('should route to 404 page if data not found in server', () => {
        // GIVEN
        spyOn(service, 'find').and.returnValue(of(new HttpResponse({ body: null })));
        mockActivatedRouteSnapshot.params = { id: 123 };

        // WHEN
        routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
          resultSach = result;
        });

        // THEN
        expect(service.find).toBeCalledWith(123);
        expect(resultSach).toEqual(undefined);
        expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
      });
    });
  });
});
